requirements.txt
altgraph==0.17
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.3.4
atomicwrites==1.4.0
attrs==21.2.0
cachetools==4.2.2
certifi==2021.10.8
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.2
Django==2.2.16
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
Faker==8.12.1
filelock==3.0.12
flake8==4.0.1
flake8-broken-line==0.3.0
flake8-isort==4.0.0
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
future==0.18.2
idna==2.9
iniconfig==1.1.1
isort==5.9.1
itypes==1.2.0
Jinja2==3.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
numpy==1.21.2
oauthlib==3.1.1
packaging==21.0
pandas==1.3.3
pdf2image==1.16.0
pefile==2019.4.18
pep8-naming==0.11.1
Pillow==8.3.1
pluggy==0.13.1
psycopg2-binary==2.9.1
py==1.10.0
pycodestyle==2.8.0
pycparser==2.20
pyflakes==2.4.0
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.10
PyJWT==2.1.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-dotenv==0.13.0
python-telegram-bot==12.7
python3-openid==3.2.0
pytils==0.3
pytz==2021.3
pywin32-ctypes==0.2.0
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
testfixtures==6.17.1
text-unidecode==1.3
toml==0.10.2
tornado==6.1
typing-extensions==3.10.0.0
tzdata==2021.2.post0
tzlocal==3.0
uritemplate==4.0.0
urllib3==1.25.9
virtualenv==20.4.7
